feat: add modal wrapper and forms
